currency module corrections

+++ src/main/java/com/training/demo/service/CoinManagementService.java
@@ -14,13 +14,28 @@ public class CoinManagementService {
 
 	public String addAllCoin(CoinManagement coinManagement) {
 			CoinManagement tempCoinManagement;
+			String coinName = coinManagement.getCoinName();
+			Double price = coinManagement.getPrice();
+			String convertedPrice = Double.toString(price);
+			if(coinName.equals("") || coinName.isEmpty() || coinName == null || coinName.trim().length() == 0)
+			{
+				return  "Coin Name Can Not Be Null Or Empty.";
+			}
+			if(convertedPrice.equals("") || price == null || convertedPrice.isEmpty() || convertedPrice.trim().length() == 0 ||convertedPrice.equals("0"))
+			{
+				return  "price Can Not Be Null Or Empty or 0.";
+			}
+			if(price<0)
+			{
+				return "Price Can Not Be Negative.";
+			}
 			tempCoinManagement = coinManagementRepository.findOneByCoinName(coinManagement.getCoinName());			
-			if(tempCoinManagement == null){
-			//Double coinInInr = coinManagement.getPrice() * coinManagement.getInitialSupply();
-			coinManagement.setProfit(0.0);
-			coinManagement.setCoinInINR(0.0);
-			coinManagementRepository.save(coinManagement);
-			return "Your Coin Has Been Added Successfully.";
+			if(tempCoinManagement == null)
+			{
+				coinManagement.setProfit(0.0);
+				coinManagement.setCoinInINR(0.0);
+				coinManagementRepository.save(coinManagement);
+				return "Your Coin Has Been Added Successfully.";
 			}
 			else
 			{
@@ -32,19 +47,48 @@ public class CoinManagementService {
 		return coinManagementRepository.findAll();
 	}
 
-	public String update(CoinManagement data) {
+	public String update(CoinManagement coinManagement) {
 		CoinManagement coinManagementData=null;		
-		coinManagementData=coinManagementRepository.findOneByCoinId(data.getCoinId());
-		coinManagementData.setCoinName(data.getCoinName());
-		coinManagementData.setSymbol(data.getSymbol());
-		coinManagementData.setInitialSupply(data.getInitialSupply());
-		coinManagementData.setPrice(data.getPrice());	
+		String coinName = coinManagement.getCoinName();
+		Double price = coinManagement.getPrice();
+		String convertedPrice = Double.toString(price);
+		if(coinName.equals("") || coinName.isEmpty() || coinName == null || coinName.trim().length() == 0)
+		{
+			return  "Coin Name Can Not Be Null Or Empty.";
+		}
+		if(convertedPrice.equals("") || price == null || convertedPrice.isEmpty() || convertedPrice.trim().length() == 0 ||convertedPrice.equals("0"))
+		{
+			return  "price Can Not Be Null Or Empty or 0.";
+		}
+		if(price<0)
+		{
+			return "Price Can Not Be Negative.";
+		}
+		
+		
+		try {
+		coinManagementData=coinManagementRepository.findOneByCoinId(coinManagement.getCoinId());
+		}
+		catch(Exception e)
+		{
+			return "invalid coinId.";
+		}
+		coinManagementData.setCoinName(coinManagement.getCoinName());
+		coinManagementData.setSymbol(coinManagement.getSymbol());
+		coinManagementData.setInitialSupply(coinManagement.getInitialSupply());
+		coinManagementData.setPrice(coinManagement.getPrice());	
+		coinManagementData.setExchangeRate(coinManagement.getExchangeRate());
+		coinManagementData.setFees(coinManagement.getFees());
 		coinManagementData=coinManagementRepository.save(coinManagementData);
 		return "Your Coin Has Been Updated Successfully.";
 	}
 
 	public String delete(Integer id) {
-		
+		String convertedId = id.toString();
+		if(convertedId.equals("") || convertedId.isEmpty() || convertedId == null || convertedId.trim().length() == 0 || convertedId.equals("0"))
+		{
+			return  "Coin Name Can Not Be Null Or Empty Or Zero.";
+		}
 		 coinManagementRepository.deleteById(id);
 		 return "Your Coin Has Been Deleted Successfully.";
 	}
